requirements.txt
# Requirements file
# Usage: pip install -r requirements.txt

torch==1.12.0
torchvision==0.13
pytorch-lightning==1.7.0
numpy==1.21.6
pillow==9.2.0
tqdm==4.64.0
matplotlib==3.5.1
pandas==1.4.2
albumentations==1.2.1
opencv-python==4.5.5
yacs==0.1.8
wandb==0.13.1

torchmetrics~=0.9.3
scikit-learn~=1.0.2
yaml~=0.2.5
pyyaml~=6.0